backend/pyproject.toml
[project]
name = "studenthunterapi"
version = "0.1.0"
description = ""
authors = [
    {name = "studenthunter",email = "info@studenthunter.kz"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
